+++ SP1Framework/Money.h
@@ -2,15 +2,15 @@
 class Money
 {
 protected:
-	int MoneyCount;
+	double MoneyCount;
 
 public:
 	Money();
 	~Money();
 	int setMoney(int);
 	bool spendMoney(int price);
-	void receiveMoney(int amt);
-	int getMoney();
+	void receiveMoney(double amt);
+	double getMoney();
 
 };
